requirements.txt
ArchivesSnake==0.6.0
atomicwrites==1.1.5
attrs==18.1.0
boltons==18.0.0
certifi==2018.8.13
chardet==3.0.4
clinner==1.12.3
colorlog==3.1.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.1
Django==2.0
django-common-helpers==0.9.2
django-cron==0.5.0
djangorestframework==3.8.2
drf-yasg==1.7.4
ElectronBonder==0.3.1
future==0.16.0
gitdb2==2.0.4
GitPython==2.1.11
health-check==3.4.1
idna==2.7
inflection==0.3.1
iso8601==0.1.12
isodate==0.6.0
itypes==1.1.0
Jinja2==2.10
jsonschema==2.6.0
MarkupSafe==1.0
more-itertools==4.3.0
openapi-codec==1.3.2
pluggy==0.7.1
psycopg2-binary==2.7.4
py==1.5.4
pycountry==18.5.26
pyfc4==0.1
pyparsing==2.2.0
pytest==3.7.1
pytest-cov==2.5.1
pytz==2018.5
PyYAML==3.13
rdflib==4.2.2
rdflib-jsonld==0.4.0
requests==2.19.1
ruamel.yaml==0.15.54
six==1.11.0
smmap2==2.0.4
structlog==18.1.0
swagger-spec-validator==2.3.1
uritemplate==3.0.0
urllib3==1.23
uuid==1.30
